make V using gcc to apply optimizations from gcc