@jinyus 2023-10-20
Main branch:
```
Vlang |    48.90 ms    |    742.24 ms    |    6.59 s
```

This PR:
```
Vlang:

        Benchmark 1: ./related
        Processing time (w/o IO): 48.892ms
        Processing time (w/o IO): 49.005ms
        Processing time (w/o IO): 49.078ms
        Processing time (w/o IO): 49.172ms
        Processing time (w/o IO): 48.492ms
        Processing time (w/o IO): 48.570ms
        Processing time (w/o IO): 49.014ms
        Processing time (w/o IO): 50.572ms
        Processing time (w/o IO): 48.341ms
        Processing time (w/o IO): 49.368ms
        Processing time (w/o IO): 48.935ms
        Processing time (w/o IO): 49.188ms
        Processing time (w/o IO): 49.251ms
          Time (mean ± σ):     389.7 ms ±   3.2 ms    [User: 361.3 ms, System: 28.2 ms]
          Range (min … max):   384.8 ms … 395.0 ms    10 runs
         
Vlang:

        Benchmark 1: ./related
        Processing time (w/o IO): 741.807ms
        Processing time (w/o IO): 740.914ms
        Processing time (w/o IO): 741.157ms
          Time (mean ± σ):     10.982 s ±  0.029 s    [User: 10.839 s, System: 0.140 s]
          Range (min … max):   10.962 s … 11.003 s    2 runs
         
Vlang:

        Benchmark 1: ./related
        Processing time (w/o IO): 6.592s
        Processing time (w/o IO): 6.595s
        Processing time (w/o IO): 6.590s
          Time (mean ± σ):     103.112 s ±  0.611 s    [User: 102.711 s, System: 0.387 s]
          Range (min … max):   102.680 s … 103.544 s    2 runs

```